Fix SDRAM mode setting procedure
According to the hardware manual, you need to do a 16-bit write, not a
32-bit write, when setting the mode in the external SDRAM chip.

+++ board/renesas/genmai/genmai.c
@@ -279,8 +279,8 @@ int board_early_init_f(void)
 	/* Check the hardware manual if your settings differ */
 	#define SDRAM_MODE_CS2 0x3FFFD040
 	#define SDRAM_MODE_CS3 0x3FFFE040
-	*(u32 *)SDRAM_MODE_CS2 = 0;
-	*(u32 *)SDRAM_MODE_CS3 = 0;
+	*(u16 *)SDRAM_MODE_CS2 = 0;
+	*(u16 *)SDRAM_MODE_CS3 = 0;
 
 	return 0;
 }
